backend/requirements.txt
django 4.2
django-filter 24.2 
djangorestframework 3.15.2
markdown 3.6
django-cors-headers 4.4.0
djangorestframework-simplejwt 5.3.1 
pyjwt 2.8.0
Pillow 10.4.0














mailtrap-2.0.1